refactor: decompose query cli execution
Split the query-first CLI into typed execution planning, action helpers,
and output helpers so selection, mutation, streaming, and formatting no
longer live in one interpreter module.
